+++ FashionToday_Back/src/main/resources/application.properties
@@ -5,3 +5,6 @@ spring.jpa.properties.hibernate.show_sql=true
 spring.jpa.hibernate.ddl-auto=update
 spring.datasource.sql-script-encoding=UTF-8
 spring.datasource.continue-on-error=true
+
+spring.mvc.view.prefix=/WEB-INF/views/
+spring.mvc.view.suffix=.jsp
