+++ FashionToday_Back/src/main/webapp/WEB-INF/views/index.jsp
@@ -1,0 +1,20 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+         pageEncoding="UTF-8"%>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
+<!DOCTYPE html>
+<html>
+<head>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+    <title>Insert title here</title>
+</head>
+<body>
+<c:if test="${userId eq null}">
+    <a href="https://kauth.kakao.com/oauth/authorize?client_id=ac9e9659b36eef40d08466896116a84a&redirect_uri=http://localhost:8080/login&response_type=code">
+        <img src="/img/kakao_account_login_btn_medium_wide_ov.png">
+    </a>
+</c:if>
+<c:if test="${userId ne null}">
+    <h1>로그인 성공입니다</h1>
+</c:if>
+</body>
+</html>
